requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.3.1
numpy>=1.21.1
opencv-python>=4.5.3
scikit-learn>=0.24.2
matplotlib>=3.4.2
pandas>=1.3.1
tensorflow>=2.6.0
keras>=2.6.0
efficientnet>=1.0.0
gradio>=2.0.0
requests>=2.26.0
uvicorn>=0.14.0
fastapi>=0.68.0
python-multipart>=0.0.5
aiofiles>=0.7.0
tqdm>=4.62.0
pytest>=6.2.4
python-dotenv>=0.19.0
scipy>=1.7.1
scikit-image>=0.18.2
jupyter>=1.0.0
notebook>=6.4.3
ipykernel>=6.0.3